- all we lack now is for the passes to match on sort in all cases

+++ Assign_1/Assign_1/real.h
@@ -16,9 +16,9 @@ typedef struct real
 	double value;
 } real;
 
-real *newReal(double);
-double getReal(real *);
-double setReal(real *, double);
+real *newReal(float);
+float getReal(real *);
+float setReal(real *, float);
 void displayReal(FILE *, void *);
 int compareReal(void *, void *);
 void freeReal(real *);
